+++ delivery/src/main/java/com/berry/delivery/application/service/delivery/DeliveryService.java
@@ -4,13 +4,13 @@ import com.berry.common.exceptionhandler.CustomApiException;
 import com.berry.common.response.ResErrorCode;
 import com.berry.common.role.RoleCheck;
 import com.berry.delivery.application.service.producer.DeliveryStatusProducer;
-import com.berry.delivery.domain.model.Delivery;
-import com.berry.delivery.domain.model.DeliveryStatus;
-import com.berry.delivery.domain.repository.DeliveryRepository;
+import com.berry.delivery.domain.model.delivery.Delivery;
+import com.berry.delivery.domain.model.delivery.DeliveryStatus;
+import com.berry.delivery.domain.repository.delivery.DeliveryRepository;
 import com.berry.delivery.presentation.dto.DeliveryDto;
-import com.berry.delivery.presentation.dto.request.DeliveryCreateRequest;
-import com.berry.delivery.presentation.dto.request.DeliveryUpdateRequest;
-import com.berry.delivery.presentation.dto.response.DeliverySearchResponse;
+import com.berry.delivery.presentation.dto.request.delivery.DeliveryCreateRequest;
+import com.berry.delivery.presentation.dto.request.delivery.DeliveryUpdateRequest;
+import com.berry.delivery.presentation.dto.response.delivery.DeliverySearchResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
