+++ delivery/src/main/java/com/berry/delivery/domain/model/delivery/DeliveryStatus.java
@@ -1,4 +1,4 @@
-package com.berry.delivery.domain.model;
+package com.berry.delivery.domain.model.delivery;
 
 public enum DeliveryStatus {
     READY,
