+++ delivery/src/main/java/com/berry/delivery/domain/repository/delivery/DeliveryRepository.java
@@ -1,10 +1,7 @@
-package com.berry.delivery.domain.repository;
+package com.berry.delivery.domain.repository.delivery;
 
 import java.util.Optional;
-import com.berry.delivery.domain.model.Delivery;
-import com.berry.delivery.domain.model.DeliveryStatus;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import com.berry.delivery.domain.model.delivery.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
