+++ delivery/src/main/java/com/berry/delivery/presentation/dto/DeliveryDto.java
@@ -1,7 +1,7 @@
 package com.berry.delivery.presentation.dto;
 
-import com.berry.delivery.domain.model.Delivery;
-import com.berry.delivery.domain.model.DeliveryStatus;
+import com.berry.delivery.domain.model.delivery.Delivery;
+import com.berry.delivery.domain.model.delivery.DeliveryStatus;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
