--- delivery/src/main/java/com/berry/delivery/presentation/dto/request/DeliveryCreateRequest.java
@@ -1,12 +1,0 @@
-package com.berry.delivery.presentation.dto.request;
-
-import com.berry.delivery.domain.model.DeliveryStatus;
-import com.berry.delivery.presentation.dto.DeliveryDto;
-
-public record DeliveryCreateRequest(
-        Long deliveryId,
-        Long receiverId,
-        Long senderId,
-        Long bidId
-        ) {
-}
