--- delivery/src/main/java/com/berry/delivery/presentation/dto/request/DeliveryUpdateRequest.java
@@ -1,9 +1,0 @@
-package com.berry.delivery.presentation.dto.request;
-
-import com.berry.delivery.domain.model.DeliveryStatus;
-
-public record DeliveryUpdateRequest(
-        String address,
-        DeliveryStatus status
-) {
-}
