+++ delivery/src/main/java/com/berry/delivery/presentation/dto/request/delivery/DeliveryCreateRequest.java
@@ -1,0 +1,9 @@
+package com.berry.delivery.presentation.dto.request.delivery;
+
+public record DeliveryCreateRequest(
+        Long deliveryId,
+        Long receiverId,
+        Long senderId,
+        Long bidId
+        ) {
+}
