+++ delivery/src/main/java/com/berry/delivery/presentation/dto/request/delivery/DeliveryUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.berry.delivery.presentation.dto.request.delivery;
+
+import com.berry.delivery.domain.model.delivery.DeliveryStatus;
+
+public record DeliveryUpdateRequest(
+        String address,
+        DeliveryStatus status
+) {
+}
